+++ hardware/pic32/variants/Cerebot_32MX4/Board_Defs.h
@@ -254,7 +254,7 @@ const static uint8_t SCK  = 11;		// PIC32 SCK2
 extern const uint32_t	port_to_tris_PGM[];
 extern const uint8_t	digital_pin_to_port_PGM[];
 extern const uint16_t	digital_pin_to_bit_mask_PGM[];
-extern const uint8_t	digital_pin_to_timer_PGM[];
+extern const uint16_t	digital_pin_to_timer_PGM[];
 extern const uint8_t 	digital_pin_to_analog_PGM[];
 extern const uint8_t	analog_pin_to_channel_PGM[];
